Rename backend to TenshiTranslatorCLI
